test: 기존 RoundInfo 상수 생성으로 분리
