+++ src/test/kotlin/lotto/service/LottoPurchaseServiceConfirmTest.kt
@@ -4,9 +4,8 @@ import TestConstant
 import app.TestConfig
 import auth.domain.vo.AuthenticatedMember
 import config.ImplementationTest
-import lotto.domain.entity.LottoBill
+import lotto.Fixture.LottoRoundFixture.createOngoingLottoRoundInfo
 import lotto.domain.entity.LottoPublish
-import lotto.domain.repository.LottoBillRepository
 import lotto.domain.repository.LottoPublishRepository
 import lotto.domain.repository.LottoRoundInfoRepository
 import lotto.domain.vo.Currency
@@ -20,17 +19,12 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Import
-import purchase.domain.entity.Purchase
-import purchase.domain.entity.PurchaseInfo
-import purchase.domain.repository.PurchaseRepository
-import purchase.domain.vo.PaymentMethod
-import purchase.domain.vo.PurchaseProvider
 import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @ImplementationTest
 @Import(TestConfig::class)
-class LottoPurchaseServiceTest {
+class LottoPurchaseServiceConfirmTest {
 
     @Autowired
     private lateinit var lottoPurchaseService: LottoPurchaseService
@@ -41,6 +35,9 @@ class LottoPurchaseServiceTest {
     @Autowired
     private lateinit var orderRepository: OrderRepository
 
+    @Autowired
+    private lateinit var lottoPublishRepository: LottoPublishRepository
+
     private var lottoPublishId: Long = 0
 
     @Nested
@@ -48,7 +45,7 @@ class LottoPurchaseServiceTest {
         @BeforeEach
         fun setUp() {
             val roundInfo = lottoRoundInfoRepository.save(
-                TestConstant.ONGOING_LOTTO_ROUND
+                createOngoingLottoRoundInfo()
             )
             orderRepository.save(
                 Order(
@@ -76,54 +73,6 @@ class LottoPurchaseServiceTest {
                         orderId = "orderId"
                     ),
                     lottoPublishId = lottoPublishId,
-                    authenticated = AuthenticatedMember("ID", "user@email.com")
-                )
-            }
-        }
-    }
-
-    @Autowired
-    private lateinit var lottoPublishRepository: LottoPublishRepository
-
-    @Autowired
-    private lateinit var purchaseRepository: PurchaseRepository
-
-    @Autowired
-    private lateinit var lottoBillRepository: LottoBillRepository
-
-    @Nested
-    inner class CancelCase {
-        private lateinit var bill: LottoBill
-
-        @BeforeEach
-        fun setup() {
-            val publish = lottoPublishRepository.save(
-                LottoPublish(
-                    lottoRoundInfo = lottoRoundInfoRepository.save(
-                        TestConstant.ONGOING_LOTTO_ROUND
-                    ),
-                    issuedAt = TestConstant.DATE_TIME,
-                )
-            )
-            val purchase = purchaseRepository.save(
-                Purchase(
-                    paymentKey = "paymentKey",
-                    orderId = "orderId",
-                    status = "SUCCESS",
-                    purchaseProvider = PurchaseProvider.TOSS,
-                    purchaseInfo = PurchaseInfo(totalAmount = BigDecimal(1000), method = PaymentMethod.CARD)
-                )
-            )
-            bill = lottoBillRepository.save(
-                LottoBill(lottoPublish = publish, purchase = purchase, memberId = "ID")
-            )
-        }
-
-        @Test
-        fun `취소가 성공적으로 진행된다`() {
-            assertDoesNotThrow {
-                lottoPurchaseService.cancel(
-                    billId = bill.getId()!!,
                     authenticated = AuthenticatedMember("ID", "user@email.com")
                 )
             }
